# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "User-Manager service is service for managing users.",
    "title": "User-Manager"
  },
  "paths": {
...
    "/user/activation": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#activate-user",
        "tags": [
          "User"
        ],
        "summary": "Activate user.",
        "operationId": "ActivateHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Link"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user activated",
            "schema": {
              "$ref": "#/definitions/CreateTokenResponse"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/user/delete/complete": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#delete-user-from-everywhere",
        "tags": [
          "User"
        ],
        "summary": "Delete user completely (almost).",
        "operationId": "CompleteDeleteHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserAgentHeader"
          },
          {
            "$ref": "#/parameters/FingerprintHeader"
          },
          {
            "$ref": "#/parameters/ClientIPHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "user deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/user/delete/partial": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#delete-user-pseudo",
        "tags": [
          "User"
        ],
        "summary": "Mark user as deleted.",
        "operationId": "PartialDeleteHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/UserIDHeader"
          }
        ],
        "responses": {
          "202": {
            "description": "user deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/user/sign_up": {
      "post": {
        "description": "https://ch.pages.containerum.net/api-docs/modules/user-manager/index.html#create-user",
        "tags": [
          "User"
        ],
        "summary": "Create user.",
        "operationId": "UserCreateHandler",
        "parameters": [
          {
            "$ref": "#/parameters/UserAgentHeader"
          },
          {
            "$ref": "#/parameters/FingerprintHeader"
          },
          {
            "$ref": "#/parameters/ClientIPHeader"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "user created",
            "schema": {
              "$ref": "#/definitions/UserLogin"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        },
        "x-method-visibility": "public"
      }
    },
...
  },
  "definitions": {
...
    "CreateTokenResponse": {
      "type": "object",
      "title": "CreateTokenResponse contains access and refresh token.",
      "properties": {
        "access_token": {
          "type": "string",
          "x-go-name": "AccessToken"
        },
        "refresh_token": {
          "type": "string",
          "x-go-name": "RefreshToken"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/git.containerum.net/ch/auth/proto"
    },
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
...
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/user-manager/vendor/github.com/containerum/cherry"
    },
...
    "Link": {
      "description": "Link -- link (for registration/activation/etc)",
      "type": "object",
      "required": [
        "link"
      ],
      "properties": {
        "created_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "CreatedAt"
        },
        "expired_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "ExpiredAt"
        },
        "is_active": {
          "type": "boolean",
          "x-go-name": "IsActive"
        },
        "link": {
          "type": "string",
          "x-go-name": "Link"
        },
        "sent_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "SentAt"
        },
        "type": {
          "$ref": "#/definitions/LinkType"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
    "LinkType": {
      "description": "LinkType -- link type",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
    "RegisterRequest": {
      "description": "RegisterRequest -- request to create new user",
      "type": "object",
      "required": [
        "login",
        "password",
        "recaptcha"
      ],
      "properties": {
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "recaptcha": {
          "type": "string",
          "x-go-name": "ReCaptcha"
        },
        "referral": {
          "type": "string",
          "x-go-name": "Referral"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
    "UserLogin": {
      "description": "UserList -- model for user login, password and id",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "x-go-name": "ID"
        },
        "login": {
          "type": "string",
          "x-go-name": "Login"
        },
        "password": {
          "type": "string",
          "x-go-name": "Password"
        }
      },
      "x-go-package": "git.containerum.net/ch/user-manager/pkg/models"
    },
...
  },
  "parameters": {
    "ClientIPHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/ClientIPHeader"
    },
    "FingerprintHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserClientHeader"
    },
...
    "UserAgentHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserAgentHeader"
    },
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}